+++ jest.config.js
@@ -5,14 +5,12 @@ module.exports = {
   // A list of reporter names that Jest uses when writing coverage reports
   coverageReporters: ['json', 'html'],
   // An array of directory names to be searched recursively up from the requiring module's location
-  moduleDirectories: ['.', 'node_modules', 'src'],
+  moduleDirectories: ['node_modules', '<rootDir>', 'src'],
   // An array of file extensions your modules use
   moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx'],
 
   moduleNameMapper: {
-    ...pathsToModuleNameMapper(
-      compilerOptions.paths /* , { prefix: '<rootDir>/' }, */,
-    ),
+    ...pathsToModuleNameMapper(compilerOptions.paths, { prefix: '<rootDir>/' }),
   },
   // A preset that is used as a base for Jest's configuration
   preset: 'ts-jest',