+++ packages/evmosjs/src/index.ts
@@ -1,7 +1,7 @@
-import * as addressConverter from '@tharsis/address-converter'
-import * as eip712 from '@tharsis/eip712'
-import * as proto from '@tharsis/proto'
-import * as provider from '@tharsis/provider'
-import * as transactions from '@tharsis/transactions'
+import * as addressConverter from '@evmos/address-converter'
+import * as eip712 from '@evmos/eip712'
+import * as proto from '@evmos/proto'
+import * as provider from '@evmos/provider'
+import * as transactions from '@evmos/transactions'
 
 export { addressConverter, eip712, proto, provider, transactions }