+++ packages/proto/src/messages/validator.ts
@@ -1,7 +1,43 @@
-import * as staking from '../proto/cosmos/staking/v1beta1/tx'
 import * as stakingTypes from '../proto/cosmos/staking/v1beta1/staking'
+import * as staking from '../proto/cosmos/staking/v1beta1/tx'
 import * as coin from '../proto/cosmos/base/v1beta1/coin'
 import { createAnyMessage, createPubKey } from './utils'
+
+const NOT_MODIFY = '[do-not-modify]'
+
+export interface MsgEditValidatorProtoInterface {
+  path: string
+  message: staking.cosmos.staking.v1beta1.MsgEditValidator
+}
+
+export function createMsgEditValidator(
+  moniker: string | undefined,
+  identity: string | undefined,
+  website: string | undefined,
+  securityContact: string | undefined,
+  details: string | undefined,
+  validatorAddress: string | undefined,
+  commissionRate: string | undefined,
+  minSelfDelegation: string | undefined,
+) {
+  const message = new staking.cosmos.staking.v1beta1.MsgEditValidator({
+    description: new stakingTypes.cosmos.staking.v1beta1.Description({
+      moniker: moniker || NOT_MODIFY,
+      identity: identity || NOT_MODIFY,
+      website: website || NOT_MODIFY,
+      security_contact: securityContact || NOT_MODIFY,
+      details: details || NOT_MODIFY,
+    }),
+    validator_address: validatorAddress,
+    commission_rate: commissionRate,
+    min_self_delegation: minSelfDelegation,
+  })
+
+  return {
+    message,
+    path: 'cosmos.staking.v1beta1.MsgEditValidator',
+  }
+}
 
 export function createMsgCreateValidator(
   validatorDescription: {