+++ packages/evmosjs/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.10](https://github.com/evmos/evmosjs/compare/evmosjs@0.2.7...evmosjs@0.2.10) (2022-09-12)
+
+**Note:** Version bump only for package evmosjs
+
+## [0.2.9](https://github.com/evmos/evmosjs/compare/evmosjs@0.2.7...evmosjs@0.2.9) (2022-08-23)
+
+**Note:** Version bump only for package evmosjs
+
+## [0.2.8](https://github.com/evmos/evmosjs/compare/evmosjs@0.2.7...evmosjs@0.2.8) (2022-07-19)
+
+**Note:** Version bump only for package evmosjs
+
+## [0.2.7](https://github.com/evmos/evmosjs/compare/evmosjs@0.2.6...evmosjs@0.2.7) (2022-06-23)
+
+**Note:** Version bump only for package evmosjs
+
+## [0.2.6](https://github.com/tharsis/evmosjs/compare/evmosjs@0.2.5...evmosjs@0.2.6) (2022-05-29)
+
+**Note:** Version bump only for package evmosjs
+
 ## [0.2.5](https://github.com/tharsis/evmosjs/compare/evmosjs@0.2.4...evmosjs@0.2.5) (2022-03-25)
 
 **Note:** Version bump only for package evmosjs